+++ packages/aws-lambda/lambdas/bin/deploy-demo.sh
@@ -24,13 +24,16 @@ EOF
 
 REGION=us-east-2
 
-for lambda_zip_file in demo-*.zip ; do
-  echo
-  echo Found zip file: $lambda_zip_file
-  if [[ $lambda_zip_file == demo-ec2-app.zip ]]; then
-    echo Skipping $lambda_zip_file.
-    continue
+function deploy_zip {
+  lambda_zip_file=$1
+
+  if [[ ! -e $lambda_zip_file ]]; then
+    echo "Zip file $lambda_zip_file does not exist, terminating."
+    exit 1
   fi
+
+  echo
+  echo "Found zip file: $lambda_zip_file"
 
   function_name=${lambda_zip_file%.zip}
   echo Deploying zip $lambda_zip_file as function $function_name to region $REGION
@@ -47,7 +50,7 @@ for lambda_zip_file in demo-*.zip ; do
 
   echo
   if [[ $needs_layer == 0 ]]; then
-    echo "The zip file $lambda_zip_file seems to contain the package @instana/aws-lambda, so I won't add the Lambda layer to it. I\'ll check if it currently has a layer that needs to be removed."
+    echo "The zip file $lambda_zip_file seems to contain the package @instana/aws-lambda, so I won't add the Lambda layer to it. I'll check if it currently has a layer that needs to be removed."
 
     current_layers=$(aws --region $REGION lambda get-function-configuration \
         --function-name $function_name \
@@ -55,7 +58,7 @@ for lambda_zip_file in demo-*.zip ; do
         | jq ".Layers")
 
     if [[ "$current_layers" =~ ":instana:" ]]; then
-      echo This lambda function definition currently has the Instana layer configured, removing it now. I\'ll also set standard handler in case the auto-wrap handler has been configured.
+      echo "This lambda function definition currently has the Instana layer configured, removing it now. I'll also set the standard handler index.handler (just in in case the auto-wrap handler had been configured previously)."
       aws --region $REGION lambda update-function-configuration \
         --function-name $function_name \
         --layers [] \
@@ -88,22 +91,36 @@ for lambda_zip_file in demo-*.zip ; do
       if [[ "$current_layers" =~ ":instana:$LAYER_VERSION" ]]; then
         echo This lambda function definition already has the specified version of the Instana layer, doing nothing.
       else
-        echo "This lambda function definition already has the Instana layer configured but uses a version ($current_layers) that is different from the one you specified ($LAYER_VERSION). I\'ll try to replace this with the specified layer. I\'ll also set the auto-wrap handler."
+        echo "This lambda function definition already has the Instana layer configured but uses a version ($current_layers) that is different from the one you specified ($LAYER_VERSION). I'll try to replace this with the specified layer. I'll also set the auto-wrap handler."
         aws --region $REGION lambda update-function-configuration \
           --function-name $function_name \
           --layers "$LAYER_ARN" \
           --handler instana-aws-lambda-auto-wrap.handler
       fi
     else
-      echo This lambda function definition currently has no Instana layer at all, adding it now. I\'ll also set the auto-wrap handler.
+      echo "This lambda function definition currently has no Instana layer at all, adding it now. I'll also set the auto-wrap handler."
       aws --region $REGION lambda update-function-configuration \
         --function-name $function_name \
         --layers "$LAYER_ARN" \
         --handler instana-aws-lambda-auto-wrap.handler
     fi
   fi
+}
 
-done
+if [[ -z $1 ]]; then
+  echo Deploying all demo zip files.
+  echo
+  for zip_file in demo-*.zip ; do
+    if [[ $lambda_zip_file == demo-ec2-app.zip ]]; then
+      echo Skipping $lambda_zip_file.
+      continue
+    fi
+    deploy_zip $zip_file
+  done
+else
+  echo "Deploying only $1.zip"
+  deploy_zip $1.zip
+fi
 
 echo
 echo Done.